# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  title: string;
  type: string;
  venue: string;
  location: string;
  /** @nullable */
  description?: string | null;
  energy_score: number;
  attendee_count: number;
  starts_at: string;
  /** @nullable */
  ends_at?: string | null;
  color_theme: string;
  joined: boolean;
  /** @nullable */
  distance_km?: number | null;
}

export interface EventInput {
  title: string;
  type: string;
  venue: string;
  location: string;
  description?: string;
  starts_at: string;
  ends_at?: string;
}

export interface EnergyScore {
  event_id: number;
  score: number;
  attendee_count: number;
  updated_at?: string;
}

export interface TimelineItem {
  id: number;
  label: string;
  description: string;
  time: string;
}

export interface Memory {
  id: number;
  title: string;
  energy_score: number;
  night_date: string;
  moments_count: number;
  vibe_matches: number;
  status: string;
  timeline?: TimelineItem[];
  /** @nullable */
  event_id?: number | null;
}

export interface MemoryInput {
  title: string;
  night_date: string;
  event_id?: number;
}

export interface Invite {
  id: number;
  event_id: number;
  event_title: string;
  sender_name: string;
  status: string;
  starts_at: string;
  attendee_count: number;
  /** @nullable */
  message?: string | null;
}

export interface InviteInput {
  event_id: number;
  recipient_username: string;
  message?: string;
}

export interface Message {
  id: number;
  content: string;
  sender_name: string;
  sent_at: string;
  is_own: boolean;
  /** @nullable */
  event_id?: number | null;
}

export interface MessageInput {
  content: string;
  event_id?: number;
}

export interface User {
  id: number;
  username: string;
  bio: string;
  nights_count: number;
  memories_count: number;
  circles_count: number;
  vibe_label: string;
  avatar_gradient?: string;
}

export interface UserUpdate {
  username?: string;
  bio?: string;
  vibe_label?: string;
}

export interface FeedStats {
  live_events: number;
  people_out: number;
  avg_energy: number;
  memories_tonight: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  label: string;
  description: string;
  timestamp: string;
}

export type ListEventsParams = {
  filter?: ListEventsFilter;
};

export type ListEventsFilter =
  (typeof ListEventsFilter)[keyof typeof ListEventsFilter];

export const ListEventsFilter = {
  trending: "trending",
  tonight: "tonight",
  friends: "friends",
  all: "all",
} as const;

export type ListMessagesParams = {
  event_id?: number;
};