# lib/api-zod/src/generated/types/energyScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyScore {
  event_id: number;
  score: number;
  attendee_count: number;
  updated_at?: string;
}