# lib/api-zod/src/generated/types/feedStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface FeedStats {
  live_events: number;
  people_out: number;
  avg_energy: number;
  memories_tonight: number;
}