# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface Invite {
  id: number;
  event_id: number;
  event_title: string;
  sender_name: string;
  status: string;
  starts_at: string;
  attendee_count: number;
  /** @nullable */
  message?: string | null;
}