# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  content: string;
  sender_name: string;
  sent_at: string;
  is_own: boolean;
  /** @nullable */
  event_id?: number | null;
}