# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  bio: string;
  nights_count: number;
  memories_count: number;
  circles_count: number;
  vibe_label: string;
  avatar_gradient?: string;
}