# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nightlog API — cinematic social nightlife platform
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  username?: string;
  bio?: string;
  vibe_label?: string;
}